Debugged indexing error caused by right-sided sorting
It turns out that I needed to set 'i' equal to 'l' when running the Partition3 function.  Previously I was setting it to 0.  After getting some help from someone who already solved the problem, I realized that setting i to 0 was causing the while loop to run for too long.  The break condition is for the while loop to run so long as it is less than or equal to right index copy, but since some recursive calls would start with a l value somewhere in the middle of the array, the 'i' was incrementing long enough for left index copy value to exceed the possible number of indices available in the array.